# optimize opencl kernel building time (still has problem in win10)