now, for ssd model , in linux it cost 5s to load, and windows ~10s...
I have done the optimze for kernel build in windows, linux need add  set_exepath(); 
I have not test in linux, just for windows  (my isv need windows version) to speedup the loading time.
now, the status in windows is still bad. if kernel always build. 
the clCreateBuffer will cost 2s, and clGetplatformids 2s.  clBuildProgram  7s

after optimzation, clBuildProgram  gone,  but clCreateBuffer 9s... I think it would be driver issue...
I use Skylake GT3e, and kabylake GT2, the same in win10.  the driver is  https://downloadmirror.intel.com/27803/a08/win64_24.20.100.6094.exe