@liyuming1978 2018-06-06
in win10, I find the cpu cost is high when do infer (or wait infer-request),  the clWaitForEvents seems to a busy wait.