@liyuming1978 2018-06-12
for the long time of clCreateBuffer ,  got it,  too much clCreateBuffer  (>30000) , I will try drop 7.0 since it has memory optimization.