@liyuming1978 2018-06-22
event though I have not see the gpu driver code,  but I guess to wait in driver need wait a very quick delay, so the code just use busywait.  I guess, just a guess:).   the delay should change to ::
if time<xx:  busywait
else: wait