@MichalMrozek 2018-06-22
Hello @liyuming1978 , I am from OpenCL driver team.
Our code works as you think, this is busy wait for some time, then it switches to non busy wait.
It is configurable with following registry keys:

DECLARE_DEBUG_VARIABLE(int32_t, OverrideEnableKmdNotify, -1, "-1: dont override, 0: disable, 1: enable")
DECLARE_DEBUG_VARIABLE(int32_t, OverrideKmdNotifyDelayMicroseconds, -1, "-1: dont override, 0: infinite timeout, >0: timeout in microseconds")
DECLARE_DEBUG_VARIABLE(int32_t, OverrideEnableQuickKmdSleep, -1, "-1: dont override, 0: disable, 1: enable. It works only when Kmd Notify is enabled.")
DECLARE_DEBUG_VARIABLE(int32_t, OverrideQuickKmdSleepDelayMicroseconds, -1, "-1: dont override, 0: infinite timeout, >0: timeout in microseconds")
DECLARE_DEBUG_VARIABLE(int32_t, OverrideEnableQuickKmdSleepForSporadicWaits, -1, "-1: dont override, 0: disable, 1: enable. It works only when QuickKmdSleep is enabled.")
DECLARE_DEBUG_VARIABLE(int32_t, OverrideDelayQuickKmdSleepForSporadicWaitsMicroseconds, -1, "-1: dont override, >0: timeout in microseconds")

It is also platform dependent, so different SKUs have different timers.
Looks like in your case the wait is not long enough to hit the non busy waiting loop.
If you are interested in building the OpenCL driver I can provide more deep explanation how to set those fields. 

But be warned, this may decrease performance, due to completion latencies in non busy mode.

For long compilation time, we have an experimental feature in the driver called cl_cache.
It is not production quality yet ( use with caution ) , but if you are willing full to try it is very easy to do so.
Just create cl_cache directory in the directory where you execute the app.
It should be automatically populated with binaries which will be re-used in further iterations.