@liyuming1978 2018-06-22
@MichalMrozek cl_cache works! that good to my patch.

for DECLARE_DEBUG_VARIABLE...  dose it need re-build opencl driver?  all just set register key? I use quick and small model to get best performance.  and I notice it only happens in windows, not happen in ubuntu.