@MichalMrozek 2018-06-22
Windows is a bit different.
There non busy wait is only when you are on a battery. When driver is on AC then it never goes to non busy wait ( due to performance constrains ), there is a huge penalty on Windows to go to non busy wait.

Unfortunately Windows builds are not available from open source code so you will not be able to check those flags out and registry flags do not work with official drivers as they are disabled there.