@liyuming1978 2018-06-22
@MichalMrozek  how to enable cl_cache in ubuntu?