@MichalMrozek 2018-06-22
cl_cache works the same way on Linux & Windows.
For Windows and non busy waits, I suggest to use throttle hints to create a command queue with CL_QUEUE_THROTTLE_LOW. This would inform the driver that power needs to be saved here and we will always initiate non busy wait in the driver.